removed the 300ms click delay for mobile devices.
according to Google : "Mobile browsers will wait approximately 300ms from the time that you tap the button to fire the click event. The reason for this is that the browser is waiting to see if you are actually performing a double tap."

I figured we wanted our UI to react ASAP so I fixed it by added an open source javascript packadge and binding click functions to it.